fix: change amount type from INTEGER to REAL in expense model and migration

+++ internal/repository/expense.sql.go
@@ -94,15 +94,15 @@ RETURNING
 `
 
 type CreateExpenseParams struct {
-	ID              string `json:"id"`
-	BookID          string `json:"book_id"`
-	CategoryID      string `json:"category_id"`
-	PaymentMethodID string `json:"payment_method_id"`
-	Date            int64  `json:"date"`
-	Amount          int64  `json:"amount"`
-	Remark          string `json:"remark"`
-	CreatedAt       int64  `json:"created_at"`
-	UpdatedAt       int64  `json:"updated_at"`
+	ID              string  `json:"id"`
+	BookID          string  `json:"book_id"`
+	CategoryID      string  `json:"category_id"`
+	PaymentMethodID string  `json:"payment_method_id"`
+	Date            int64   `json:"date"`
+	Amount          float64 `json:"amount"`
+	Remark          string  `json:"remark"`
+	CreatedAt       int64   `json:"created_at"`
+	UpdatedAt       int64   `json:"updated_at"`
 }
 
 func (q *Queries) CreateExpense(ctx context.Context, arg CreateExpenseParams) (int64, error) {
@@ -250,13 +250,13 @@ WHERE
 `
 
 type UpdateExpenseByIDParams struct {
-	CategoryID      string `json:"category_id"`
-	PaymentMethodID string `json:"payment_method_id"`
-	Date            int64  `json:"date"`
-	Amount          int64  `json:"amount"`
-	Remark          string `json:"remark"`
-	UpdatedAt       int64  `json:"updated_at"`
-	ID              string `json:"id"`
+	CategoryID      string  `json:"category_id"`
+	PaymentMethodID string  `json:"payment_method_id"`
+	Date            int64   `json:"date"`
+	Amount          float64 `json:"amount"`
+	Remark          string  `json:"remark"`
+	UpdatedAt       int64   `json:"updated_at"`
+	ID              string  `json:"id"`
 }
 
 func (q *Queries) UpdateExpenseByID(ctx context.Context, arg UpdateExpenseByIDParams) (int64, error) {
